+++ storybook/stories/staking/Staking.stories.js
@@ -136,4 +136,11 @@ storiesOf('Decentralization | Wizard', module)
   .addDecorator(decorator)
   .add('Delegation Wizard', props => <StakingDelegationSteps {...props} />, {
     id: 'wizard',
-  });
+  })
+  .add(
+    'Delegation Wizard - Delegation Not Available',
+    props => <StakingDelegationSteps {...props} isDisabled />,
+    {
+      id: 'wizard',
+    }
+  );